requirements.txt

psycopg2==2.8.6
requests
django-webpack-loader==0.7.0
python-dateutil
numpy
pandas
pyyaml
